correct btree ubsan issue, basically (val << 1) on a signed int64_t is undefined behavior when val is negative. casting to uint64_t first ((uint64_t)val << 1) makes the left shift well-defined; savepoint paths uses malloc for the ops array but only initialized value when value_size > 0 thus for delete operations (value_size == 0), the value pointer was left as uninitialized garbage, this has been corrected adding a branch

+++ src/btree.c
@@ -532,9 +532,9 @@ static int btree_leaf_serialize(const btree_pending_leaf_t *leaf, const int64_t 
         est_size += btree_varint_size(leaf->entries[i].value_size);
         est_size += btree_varint_size(leaf->entries[i].vlog_offset);
         const int64_t seq_delta = (int64_t)(leaf->entries[i].seq - base_seq);
-        est_size += btree_varint_size((uint64_t)((seq_delta << 1) ^ (seq_delta >> 63)));
-        est_size += btree_varint_size(
-            (uint64_t)((leaf->entries[i].ttl << 1) ^ (leaf->entries[i].ttl >> 63)));
+        est_size += btree_varint_size(((uint64_t)seq_delta << 1) ^ (uint64_t)(seq_delta >> 63));
+        est_size += btree_varint_size(((uint64_t)leaf->entries[i].ttl << 1) ^
+                                      (uint64_t)(leaf->entries[i].ttl >> 63));
         est_size += 1; /* flags */
         keys_total += suffix_lens[i];
         if (leaf->entries[i].vlog_offset == 0 && leaf->values[i])
